Build different `Status` enums for N-API v8 and v9
